Worked through a bunch
Fixed up the backend for the user tracking so that it now correctly displays just one user and the history is ordered by date. Added in collapsing functionality to keep the information clean and easy to read. Also fixed some portions of the home page so it should correctly route, maybe. May need to test, but just make sure the site is running. Everything looks in order on my side.